super-app: rename services for idiomaticness
The pathname `services/Foo` should be read as 'the Foo service' in plain
English, e.g. 'Language service', 'Config service'.

Rename the services that don't fit this pattern (e.g. 'DetailServices').
Unfortunately, we've referred to partners in plain English as Support Services,
so to avoid 'Support Service service', I've opted to use 'Support' as a
shorthand for 'Support Service'.
